cut boost dependency

+++ demo/morpion/hom/const.cpp
@@ -5,6 +5,6 @@
 
 // Const variable
 size_t NBCELL ;
-size_t STATE_SYSTEM_CELL ;
+int STATE_SYSTEM_CELL ;
 size_t LINE ;
 size_t COLUMN ;
